# jrag-model/src/main/resources/openapi-interface.yaml
openapi: 3.0.3
info:
  title: ai operate - OpenAPI 3.0
  description: |-
    This is ai operate OpenAPI.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:30110
    description: 本机调试环境
...
paths:
...
  /v1/rest/jrag/file:
    post:
      tags: [ File ]
      summary: 上传文件
      description: 上传文件
      operationId: uploadFIle
      requestBody:
        description: 上传的文件
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: './openapi-model.yaml#/components/schemas/FileDto'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server
    put:
      tags: [ File ]
      summary: 替换指定ID的文件
      description: 替换指定ID的文件
      operationId: putFile
      requestBody:
        description: 上传的文件
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: 文件ID
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: './openapi-model.yaml#/components/schemas/FileDto'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method Not Allowed
        "500":
          description: Internal Server
    delete:
      tags: [ File ]
      summary: 删除指定ID的文件
      description: 删除指定ID的文件
      operationId: deleteFile
      parameters:
        - in: query
          name: file-id
          schema:
            type: array
            items:
              type: string
              description: 文件ID
      responses:
        "200":
          description: Successful operation
  /v1/rest/jrag/file/{file-id}:
    get:
      tags: [ File ]
      operationId: getFile
      parameters:
        - in: path
          name: file-id
          schema:
            type: string
          required: true
      responses:
        200:
          description: ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        404:
          description: not found
...